+++ tests/Google/ClientTest.php
@@ -24,7 +24,9 @@ use Google\Client;
 use Google\Service\Drive;
 use Google\AuthHandler\AuthHandlerFactory;
 use Google\Auth\FetchAuthTokenCache;
+use Google\Auth\CredentialsLoader;
 use Google\Auth\GCECache;
+use Google\Auth\Credentials\GCECredentials;
 use GuzzleHttp\Client as GuzzleClient;
 use GuzzleHttp\Psr7\Request;
 use GuzzleHttp\Psr7\Response;
@@ -37,6 +39,7 @@ use ReflectionClass;
 use ReflectionMethod;
 use InvalidArgumentException;
 use Exception;
+use DomainException;
 
 class ClientTest extends BaseTest
 {
@@ -689,11 +692,20 @@ class ClientTest extends BaseTest
         $mockCacheItem->get()
             ->shouldBeCalledTimes(1)
             ->willReturn(true);
+        $mockUniverseDomainCacheItem = $this->prophesize(CacheItemInterface::class);
+        $mockUniverseDomainCacheItem->isHit()
+                ->willReturn(true);
+        $mockUniverseDomainCacheItem->get()
+            ->shouldBeCalledTimes(1)
+            ->willReturn('googleapis.com');
 
         $mockCache = $this->prophesize(CacheItemPoolInterface::class);
         $mockCache->getItem($prefix . GCECache::GCE_CACHE_KEY)
             ->shouldBeCalledTimes(1)
             ->willReturn($mockCacheItem->reveal());
+        $mockCache->getItem(GCECredentials::cacheKey . 'universe_domain')
+            ->shouldBeCalledTimes(1)
+            ->willReturn($mockUniverseDomainCacheItem->reveal());
 
         $client = new Client(['cache_config' => $cacheConfig]);
         $client->setCache($mockCache->reveal());
@@ -849,11 +861,16 @@ class ClientTest extends BaseTest
         $credentials = $this->prophesize('Google\Auth\CredentialsLoader');
         $credentials->getCacheKey()
             ->willReturn('cache-key');
+        $credentials->getUniverseDomain()
+            ->willReturn('googleapis.com');
 
         // Ensure the access token provided by our credentials loader is used
-        $credentials->fetchAuthToken(Argument::any())
+        $credentials->updateMetadata([], null, Argument::any())
             ->shouldBeCalledOnce()
-            ->willReturn(['access_token' => 'abc']);
+            ->willReturn(['authorization' => 'Bearer abc']);
+        $credentials->getLastReceivedToken()
+            ->shouldBeCalledTimes(2)
+            ->willReturn(null);
 
         $client = new Client(['credentials' => $credentials->reveal()]);
 
@@ -909,5 +926,22 @@ class ClientTest extends BaseTest
             'enable_serial_consent' => 'true'
         ]);
         $this->assertStringContainsString('&enable_serial_consent=true', $authUrl1);
+    }
+    public function testUniverseDomainMismatch()
+    {
+        $this->expectException(DomainException::class);
+        $this->expectExceptionMessage(
+            'The configured universe domain (example.com) does not match the credential universe domain (foo.com)'
+        );
+
+        $credentials = $this->prophesize(CredentialsLoader::class);
+        $credentials->getUniverseDomain()
+            ->shouldBeCalledOnce()
+            ->willReturn('foo.com');
+        $client = new Client([
+            'universe_domain' => 'example.com',
+            'credentials' => $credentials->reveal(),
+        ]);
+        $client->authorize();
     }
 }