+++ tests/Google/Service/ResourceTest.php
@@ -35,10 +35,11 @@ use Prophecy\Argument;
 
 class TestService extends \Google\Service
 {
-    public function __construct(Client $client)
+    public function __construct(Client $client, $rootUrl = null)
     {
         parent::__construct($client);
-        $this->rootUrl = "https://test.example.com";
+        $this->rootUrl = $rootUrl ?: "https://test.example.com";
+        $this->rootUrlTemplate = $rootUrl ?: "https://test.UNIVERSE_DOMAIN";
         $this->servicePath = "";
         $this->version = "v1beta1";
         $this->serviceName = "test";
@@ -59,6 +60,7 @@ class ResourceTest extends BaseTest
         $this->client->getLogger()->willReturn($logger->reveal());
         $this->client->shouldDefer()->willReturn(true);
         $this->client->getHttpClient()->willReturn(new GuzzleClient());
+        $this->client->getUniverseDomain()->willReturn('example.com');
 
         $this->service = new TestService($this->client->reveal());
     }
@@ -104,6 +106,38 @@ class ResourceTest extends BaseTest
         $this->assertEquals("https://test.example.com/method/path", (string) $request->getUri());
         $this->assertEquals("POST", $request->getMethod());
         $this->assertFalse($request->hasHeader('Content-Type'));
+        $this->assertFalse($request->hasHeader('X-Goog-Api-Version'));
+    }
+
+    public function testCallWithUniverseDomainTemplate()
+    {
+        $client = $this->prophesize(Client::class);
+        $logger = $this->prophesize("Monolog\Logger");
+        $this->client->getLogger()->willReturn($logger->reveal());
+        $this->client->shouldDefer()->willReturn(true);
+        $this->client->getHttpClient()->willReturn(new GuzzleClient());
+        $this->client->getUniverseDomain()->willReturn('example-universe-domain.com');
+
+        $this->service = new TestService($this->client->reveal());
+
+        $resource = new GoogleResource(
+            $this->service,
+            "test",
+            "testResource",
+            [
+                "methods" => [
+                    "testMethod" => [
+                        "parameters" => [],
+                        "path" => "method/path",
+                        "httpMethod" => "POST",
+                    ]
+                ]
+            ]
+        );
+        $request = $resource->call("testMethod", [[]]);
+        $this->assertEquals("https://test.example-universe-domain.com/method/path", (string) $request->getUri());
+        $this->assertEquals("POST", $request->getMethod());
+        $this->assertFalse($request->hasHeader('Content-Type'));
     }
 
     public function testCallWithPostBody()
@@ -130,9 +164,9 @@ class ResourceTest extends BaseTest
 
     public function testCallServiceDefinedRoot()
     {
-        $this->service->rootUrl = "https://sample.example.com";
+        $service = new TestService($this->client->reveal(), "https://sample.example.com");
         $resource = new GoogleResource(
-            $this->service,
+            $service,
             "test",
             "testResource",
             [
@@ -441,4 +475,32 @@ class ResourceTest extends BaseTest
             $this->assertEquals($errors, $e->getErrors());
         }
     }
+
+    public function testVersionedResource()
+    {
+        $resource = new VersionedResource(
+            $this->service,
+            "test",
+            "testResource",
+            [
+                "methods" => [
+                    "testMethod" => [
+                        "parameters" => [],
+                        "path" => "method/path",
+                        "httpMethod" => "POST",
+                    ]
+                ]
+            ]
+        );
+        $request = $resource->call("testMethod", [['postBody' => ['foo' => 'bar']]]);
+        $this->assertEquals("https://test.example.com/method/path", (string) $request->getUri());
+        $this->assertEquals("POST", $request->getMethod());
+        $this->assertTrue($request->hasHeader('X-Goog-Api-Version'));
+        $this->assertEquals('v1_20240101', $request->getHeaderLine('X-Goog-Api-Version'));
+    }
+}
+
+class VersionedResource extends GoogleResource
+{
+    protected $apiVersion = 'v1_20240101';
 }